Replace with_privileges function to separate system module
